+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/protocol/share_rest_client.hpp
@@ -772,7 +772,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       Storage::Metadata Metadata;
       int64_t ContentLength = int64_t();
       ShareFileHttpHeaders HttpHeaders;
-      Azure::Core::Nullable<std::string> ContentRange;
+      Azure::Core::Http::Range ContentRange;
       std::string ETag;
       Azure::Core::Nullable<Storage::ContentHash> TransactionalContentHash;
       std::string AcceptRanges;
@@ -6100,10 +6100,31 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
             result.ContentLength
                 = std::stoll(response.GetHeaders().at(Details::HeaderContentLength));
             result.HttpHeaders.ContentType = response.GetHeaders().at(Details::HeaderContentType);
-            if (response.GetHeaders().find(Details::HeaderContentRange)
-                != response.GetHeaders().end())
+
+            auto content_range_iterator = response.GetHeaders().find(Details::HeaderContentRange);
+            if (content_range_iterator != response.GetHeaders().end())
             {
-              result.ContentRange = response.GetHeaders().at(Details::HeaderContentRange);
+              const std::string& content_range = content_range_iterator->second;
+              auto bytes_pos = content_range.find("bytes ");
+              auto dash_pos = content_range.find("-", bytes_pos + 6);
+              auto slash_pos = content_range.find("/", dash_pos + 1);
+              int64_t range_start_offset = std::stoll(std::string(
+                  content_range.begin() + bytes_pos + 6, content_range.begin() + dash_pos));
+              int64_t range_end_offset = std::stoll(std::string(
+                  content_range.begin() + dash_pos + 1, content_range.begin() + slash_pos));
+              result.ContentRange = Azure::Core::Http::Range{
+                  range_start_offset, range_end_offset - range_start_offset + 1};
+            }
+            else
+            {
+              result.ContentRange = Azure::Core::Http::Range{
+                  0, std::stoll(response.GetHeaders().at(Details::HeaderContentLength))};
+            }
+            if (content_range_iterator != response.GetHeaders().end())
+            {
+              const std::string& content_range = content_range_iterator->second;
+              auto slash_pos = content_range.find("/");
+              result.ContentLength = std::stoll(content_range.substr(slash_pos + 1));
             }
             result.ETag = response.GetHeaders().at(Details::HeaderETag);
             if (response.GetHeaders().find(Details::HeaderTransactionalContentHashMd5)
@@ -6236,10 +6257,31 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
             result.ContentLength
                 = std::stoll(response.GetHeaders().at(Details::HeaderContentLength));
             result.HttpHeaders.ContentType = response.GetHeaders().at(Details::HeaderContentType);
-            if (response.GetHeaders().find(Details::HeaderContentRange)
-                != response.GetHeaders().end())
+
+            auto content_range_iterator = response.GetHeaders().find(Details::HeaderContentRange);
+            if (content_range_iterator != response.GetHeaders().end())
             {
-              result.ContentRange = response.GetHeaders().at(Details::HeaderContentRange);
+              const std::string& content_range = content_range_iterator->second;
+              auto bytes_pos = content_range.find("bytes ");
+              auto dash_pos = content_range.find("-", bytes_pos + 6);
+              auto slash_pos = content_range.find("/", dash_pos + 1);
+              int64_t range_start_offset = std::stoll(std::string(
+                  content_range.begin() + bytes_pos + 6, content_range.begin() + dash_pos));
+              int64_t range_end_offset = std::stoll(std::string(
+                  content_range.begin() + dash_pos + 1, content_range.begin() + slash_pos));
+              result.ContentRange = Azure::Core::Http::Range{
+                  range_start_offset, range_end_offset - range_start_offset + 1};
+            }
+            else
+            {
+              result.ContentRange = Azure::Core::Http::Range{
+                  0, std::stoll(response.GetHeaders().at(Details::HeaderContentLength))};
+            }
+            if (content_range_iterator != response.GetHeaders().end())
+            {
+              const std::string& content_range = content_range_iterator->second;
+              auto slash_pos = content_range.find("/");
+              result.ContentLength = std::stoll(content_range.substr(slash_pos + 1));
             }
             result.ETag = response.GetHeaders().at(Details::HeaderETag);
             if (response.GetHeaders().find(Details::HeaderTransactionalContentHashMd5)