+++ sdk/storage/azure-storage-files-shares/src/share_file_client.cpp
@@ -641,8 +641,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     int64_t fileRangeSize;
     if (firstChunkOptions.Range.HasValue())
     {
-      fileSize = std::stoll(firstChunk->ContentRange.GetValue().substr(
-          firstChunk->ContentRange.GetValue().find('/') + 1));
+      fileSize = firstChunk->ContentLength;
       fileRangeSize = fileSize - firstChunkOffset;
       if (options.Range.GetValue().Length.HasValue())
       {
@@ -765,8 +764,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     int64_t fileRangeSize;
     if (firstChunkOptions.Range.HasValue())
     {
-      fileSize = std::stoll(firstChunk->ContentRange.GetValue().substr(
-          firstChunk->ContentRange.GetValue().find('/') + 1));
+      fileSize = firstChunk->ContentLength;
       fileRangeSize = fileSize - firstChunkOffset;
       if (options.Range.GetValue().Length.HasValue())
       {