+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 - Removed `GetDirectoryClient` and `GetFileClient` from `ShareClient`. `ShareDirectoryClient` and `ShareFileClient` now initializes with the name of the resource, not path, to indicate that no path parsing is done for the API
+- `ContentRange` in `FileDownloadResult` is now `Azure::Core::Http::Range`.
 
 ## 12.0.0-beta.6 (2020-01-14)